# backend/api/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Social Networking API",
      "description": "API for a social networking using NodeJS and MongoDB."
    },
    "host": "localhost:4000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "securityDefinitions": {
      "JWT": {
        "type": "apiKey",
        "name": "Authorization",
        "description": "Bearer ",
        "in": "header"
      }
    },
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/v1/posts": {
        "get": {
          "tags": [
            "Post"
          ],
          "description": "This endpoint gets all posts by me.",
          "parameters": [
            
          ],
          "responses": {
            "200": {
              "description": "Posts successfully obtained."
            },
            "default": {
              "description": "Unexpected error"
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        },
        "post": {
          "tags": [
            "Post"
          ],
          "description": "This endpoint likes a post.",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "type": "string",
              "description": "Post Id."
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Post successfully liked."
            },
            "404": {
              "description": "Post not found."
            }
          },
          "security": [
            {
              "JWT": []
            }
          ]
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "john10"
          },
          "password": {
            "type": "string",
            "example": "Password123456*"
          },
          "profile": {
            "$ref": "#/definitions/Profile"
          }
        },
        "required": [
          "user",
          "password"
        ]
      },
      "Login": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "john10"
          },
          "password": {
            "type": "string",
            "example": "Password123456*"
          }
        },
        "required": [
          "user",
          "password"
        ]
      },
      "Registry": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "john"
          },
          "password": {
            "type": "string",
            "example": "Password123456*"
          },
          "name": {
            "type": "string",
            "example": "John Mark"
          }
        },
        "required": [
          "user",
          "password"
        ]
      },
      "Profile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Mark"
          },
          "user": {
            "type": "string",
            "example": "john10"
          },
          "followers": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Profile"
            }
          },
          "following": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Profile"
            }
          }
        },
        "required": [
          "name",
          "user"
        ]
      },
      "Post": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "This is a title of a post..."
          },
          "description": {
            "type": "string",
            "example": "This is a description of a post..."
          },
          "profile": {
            "$ref": "#/definitions/Profile"
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Comment"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Profile"
            }
          }
        },
        "required": [
          "title",
          "description",
          "profile"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "example": "This is a description of a comment..."
          },
          "profile": {
            "$ref": "#/definitions/Profile"
          },
          "post": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Post"
            }
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Profile"
            }
          }
        },
        "required": [
          "description",
          "profile",
          "post"
        ]
      }
    }
  }